finished error checking for template entering view
