+++ app/(groupwithlayout)/aisequences/editsequence/addstep/testtemplate/page.js
@@ -4,6 +4,7 @@ import Navbar from "@/app/ui/Navbar";
 import Link from 'next/link';
 import { useImmerReducer } from 'use-immer';
 import { useState, useEffect } from 'react';
+import { saveStep } from '@/app/lib/actions'
 
 export default function TestTemplate() {
 
@@ -14,7 +15,7 @@ export default function TestTemplate() {
 
     return (
         <div className="min-w-full h-dvh">
-            <Navbar url={"AI Sequences / New Sequence / Add Step / Test"} />
+            <Navbar url={"AI Sequences / Edit Sequence / Add Step / Test"} />
             <div className="p-4">
                 <label className="block text-xl font-medium leading-6 text-gray-900">
                     {`Paste a Linkedin URL to test your the “${stepName}” step`}
@@ -26,18 +27,20 @@ export default function TestTemplate() {
                     />
                     <Link
                         type="button"
-                        href="/aisequences/newsequence"
+                        href="/aisequences/editsequence"
                         className="ml-4 whitespace-nowrap rounded-lg bg-indigo-600 px-7 py-3 text-sm text-white shadow-sm justify-center text-center hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
                     >
                         Test
                     </Link>
-                    <Link
+                    <button
                         type="button"
-                        href="/aisequences/newsequence"
+                        onClick={async (e) => {
+                            await saveStep(localStorage.getItem('sequence_name'), localStorage.getItem("step_name"), localStorage.getItem("template"), JSON.parse(localStorage.getItem("example_information_subject_line")), JSON.parse(localStorage.getItem("example_information_body")))
+                        }}
                         className="ml-4 whitespace-nowrap rounded-lg bg-green-600 px-7 py-3 text-sm text-white shadow-sm justify-center text-center hover:bg-green-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-green-600"
                     >
                         Save step
-                    </Link>
+                    </button>
                 </div>
             </div>
         </div>
